Refactor Reachability to separate RTA-specifics from general API
Prepare for experimenting with different reachability algorithms by
splitting out a general API that the rest of the compiler can use to
determine the reachability of various program constructs from the
implementation of the RTA algorithm its internal data structures.
