+++ plugins/dispatch/src/main/java/org/qbicc/plugin/dispatch/DispatchTableEmitter.java
@@ -3,7 +3,7 @@ package org.qbicc.plugin.dispatch;
 import java.util.function.Consumer;
 
 import org.qbicc.context.CompilationContext;
-import org.qbicc.plugin.reachability.RTAInfo;
+import org.qbicc.plugin.reachability.ReachabilityInfo;
 import org.qbicc.context.ClassContext;
 import org.qbicc.type.definition.DefinedTypeDefinition;
 import org.qbicc.type.definition.LoadedTypeDefinition;
@@ -12,7 +12,7 @@ public class DispatchTableEmitter implements Consumer<CompilationContext>  {
 
     @Override
     public void accept(CompilationContext ctxt) {
-        RTAInfo info = RTAInfo.get(ctxt);
+        ReachabilityInfo info = ReachabilityInfo.get(ctxt);
         DispatchTables tables = DispatchTables.get(ctxt);
 
         // Walk down the live class hierarchy and emit vtables for each class
